chore: extract out post widget
