+++ packages/app/lib/src/app/wrapper_page.dart
@@ -30,6 +30,7 @@ class WrapperPage extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    print('wrapper page is building');
     return AutoTabsScaffold(
       routes: const [
         LocalFeedRoute(),
