+++ packages/app/lib/src/features/home/presentation/home/feed.dart
@@ -1,6 +1,4 @@
 /// This library contains a widget that displays a feed of posts.
-// ignore_for_file: prefer_expression_function_bodies
-
 library;
 
 import 'package:flutter/foundation.dart';
@@ -10,6 +8,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import '../../application/feed_service.dart';
 import '../../domain/feed_entity.dart';
 import '../../domain/post_entity.dart';
+import 'post.dart';
 
 /// {@template our_democracy.features.home.presentation.home.feed}
 /// An infinite loading list of posts.
@@ -28,7 +27,7 @@ class Feed extends ConsumerWidget {
     // Maybe change to scaffold with floating action button and list view as child
     return ListView.builder(
       shrinkWrap: true,
-      prototypeItem: _Post(
+      prototypeItem: Post(
         post: PostEntity(
           headline: '',
           description: '',
@@ -44,13 +43,13 @@ class Feed extends ConsumerWidget {
         );
 
         return switch (response) {
-          AsyncData(:final value) when value != null => _Post(post: value),
+          AsyncData(:final value) when value != null => Post(post: value),
 
           // If we run out of items, return null.
           AsyncData() => null,
 
           // If there's an error, display it as another post.
-          AsyncError(:final error) => _Post(
+          AsyncError(:final error) => Post(
               post: PostEntity(
                 headline: 'Error',
                 description: error.toString(),
@@ -73,49 +72,6 @@ class Feed extends ConsumerWidget {
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties.add(DiagnosticsProperty<FeedEntity>('feed', feed));
-  }
-  // coverage:ignore-end
-}
-
-class _Post extends StatelessWidget {
-  const _Post({
-    required this.post,
-    // Temporary ignore, see <dart-lang/sdk#49025>.
-    // ignore: unused_element
-    super.key,
-  });
-  final PostEntity post;
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: ListTile(
-        leading: switch (post.image) {
-          // If the image is an BucketFile, figure out how to get the image from the API.
-          // Will need to be cached in Riverpod. Probably store it in the entity.
-          // But, deserialization shouldn't need to know about the API.
-          // This'll be tricky. For now, I think we make a dummy image.
-          final String _ => Image.network(
-              'https://upload.wikimedia.org/wikipedia/commons/thumb/3/3f/Placeholder_view_vector.svg/991px-Placeholder_view_vector.svg.png',
-            ),
-          // final BucketFile image => Image.memory(image),
-          // Else, return null.
-          null => null,
-        },
-        title: Text(post.headline),
-        subtitle: Text(post.description),
-        // isThreeLine: true,
-        // minVerticalPadding: 100,
-        // TODO(MattsAttack): add in on tap functionality to click on post
-      ),
-    );
-  }
-
-  // coverage:ignore-start
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties.add(DiagnosticsProperty<PostEntity>('post', post));
   }
   // coverage:ignore-end
 }
